Small fixes

+++ blinktopus/src/main/java/dbse/fopj/blinktopus/api/resultmodel/LogResult.java
@@ -19,7 +19,7 @@ public class LogResult extends Result {
 	}
 
 	public LogResult(String SVid, String type, String table, String attr, double lower, double higher, long timeLog,
-			long timeSV, double exactCount, double apprCount, double error, String message, List<Tuple> resultTuples) {
+			long timeSV, long exactCount, long apprCount, double error, String message, List<Tuple> resultTuples) {
 		super(SVid, type, table, attr, lower, higher, timeLog, timeSV, exactCount, apprCount, error, message);
 		this.logResultTuples = resultTuples;
 	}
